Rename security module to "sandbox"
