+++ daemon/src/daemon.rs
@@ -2,7 +2,7 @@ use crate::connect;
 use crate::document::Document;
 use crate::editor::EditorHandle;
 use crate::ot::OTServer;
-use crate::security;
+use crate::sandbox;
 use crate::types::{
     CursorState, EditorProtocolMessageFromEditor, EditorProtocolMessageToEditor, EditorTextDelta,
     FileTextDelta, PatchEffect, Range, RevisionedEditorTextDelta, TextDelta,
@@ -102,7 +102,7 @@ impl DocumentActor {
         let persistence_file = base_dir.join(".ethersync/doc");
         let crdt_doc = if persistence_file.exists() && !init {
             info!("Loading persisted CRDT document from {persistence_file:?}");
-            let bytes = security::read_file(&base_dir, &persistence_file)
+            let bytes = sandbox::read_file(&base_dir, &persistence_file)
                 .unwrap_or_else(|_| panic!("Could not read file '{persistence_file:?}'"));
             Document::load(&bytes)
         } else {
@@ -164,7 +164,7 @@ impl DocumentActor {
                 debug!("Persisting!");
                 let bytes = self.crdt_doc.save();
                 let persistence_file = self.base_dir.join(".ethersync/doc");
-                security::write_file(&self.base_dir, &persistence_file, &bytes)
+                sandbox::write_file(&self.base_dir, &persistence_file, &bytes)
                     .unwrap_or_else(|_| panic!("Failed to persist to {persistence_file:?}"));
             }
             DocMessage::ReceiveSyncMessage {
@@ -188,7 +188,7 @@ impl DocumentActor {
                             cursor_states.push(cursor_state);
                         }
                         PatchEffect::FileRemoval(file_path) => {
-                            security::remove_file(
+                            sandbox::remove_file(
                                 &self.base_dir,
                                 Path::new(&self.absolute_path_for_file_path(&file_path)),
                             )
@@ -453,11 +453,11 @@ impl DocumentActor {
 
                 // Create the parent directorie(s), if neccessary.
                 let parent_dir = Path::new(&abs_path).parent().unwrap();
-                security::create_dir_all(&self.base_dir, parent_dir).unwrap_or_else(|_| {
+                sandbox::create_dir_all(&self.base_dir, parent_dir).unwrap_or_else(|_| {
                     panic!("Could not create parent directory {}", parent_dir.display())
                 });
 
-                security::write_file(&self.base_dir, Path::new(&abs_path), &text.into_bytes())
+                sandbox::write_file(&self.base_dir, Path::new(&abs_path), &text.into_bytes())
                     .unwrap_or_else(|_| panic!("Could not write to file {abs_path}"));
             } else {
                 warn!("Failed to get content of file '{file_path}' when writing to disk. Key should have existed?");
@@ -492,7 +492,7 @@ impl DocumentActor {
             })
             .for_each(|dir_entry| {
                 let file_path = dir_entry.path();
-                match security::read_file(&self.base_dir, file_path) {
+                match sandbox::read_file(&self.base_dir, file_path) {
                     Ok(bytes) => {
                         let relative_file_path = self.file_path_for_uri(
                             file_path
@@ -754,11 +754,11 @@ mod tests {
             let file1 = dir.child("file1");
             let file2 = dir.child("file2");
             let subdir = dir.child("sub");
-            security::create_dir(dir.path(), &subdir).unwrap();
+            sandbox::create_dir(dir.path(), &subdir).unwrap();
             let file3 = dir.child("sub/file3");
-            security::write_file(dir.path(), &file1, b"content1").unwrap();
-            security::write_file(dir.path(), &file2, b"content2").unwrap();
-            security::write_file(dir.path(), &file3, b"content3").unwrap();
+            sandbox::write_file(dir.path(), &file1, b"content1").unwrap();
+            sandbox::write_file(dir.path(), &file2, b"content2").unwrap();
+            sandbox::write_file(dir.path(), &file3, b"content3").unwrap();
             dir
         }
 
@@ -806,15 +806,15 @@ mod tests {
 
             // Thus, we only expect file1 to be changed on disk.
             assert_eq!(
-                security::read_file(dir.path(), &dir.child("file1")).unwrap(),
+                sandbox::read_file(dir.path(), &dir.child("file1")).unwrap(),
                 b"foobarcontent1",
             );
             assert_eq!(
-                security::read_file(dir.path(), &dir.child("file2")).unwrap(),
+                sandbox::read_file(dir.path(), &dir.child("file2")).unwrap(),
                 b"content2",
             );
             assert_eq!(
-                security::read_file(dir.path(), &dir.child("sub/file3")).unwrap(),
+                sandbox::read_file(dir.path(), &dir.child("sub/file3")).unwrap(),
                 b"content3",
             );
         }
